requirements.txt
Django>=3.2
djangorestframework
channels
channels_redis
daphne
python-decouple
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
celery
redis
confluent-kafka
Pillow
social-auth-app-django
google-auth
google-auth-oauthlib
google-auth-httplib2
django-allauth>=1.9
dj-rest-auth>=3.1
python-dotenv>=1.0
django-celery-beat==2.7.0
